+++ examples/hybrid/plane/density_current_2d_invariant_rhoe.jl
@@ -269,7 +269,7 @@ using Test
 end
 
 ENV["GKSwstype"] = "nul"
-import Plots, ClimaCorePlots
+import Plots
 Plots.GRBackend()
 
 dir = "dc_invariant_etot"